+++ src/Runner/Printer/Proof.php
@@ -5,6 +5,7 @@ namespace Innmind\BlackBox\Runner\Printer;
 
 use Innmind\BlackBox\Runner\{
     Proof\Scenario\Failure,
+    Assert\Debug,
     IO,
 };
 
@@ -13,6 +14,11 @@ interface Proof
     public function emptySet(IO $output, IO $error): void;
     public function success(IO $output, IO $error): void;
     public function shrunk(IO $output, IO $error): void;
-    public function failed(IO $output, IO $error, Failure $failure): void;
+    public function failed(
+        IO $output,
+        IO $error,
+        Failure $failure,
+        Debug $debug,
+    ): void;
     public function end(IO $output, IO $error): void;
 }